+++ NEWS.md
@@ -1,4 +1,4 @@
-# cli (development version)
+# cli 3.5.0
 
 * New `keypress()` function to read a single key press from a terminal.